requirements.txt
Flask==3.1.2
Flask-CORS==6.0.2
gunicorn==24.0.0
waitress==3.0.2
joblib==1.5.2
pandas==2.3.3
numpy==2.3.4
PyMuPDF==1.26.6
pdfplumber==0.11.7
pdfminer.six==20250506
PyPDF2==3.0.1
pillow==12.0.0
python-dotenv==1.2.1
cryptography==46.0.3
werkzeug==3.1.3
scikit-learn==1.7.2
openpyxl==3.1.2
sentence-transformers==5.1.2
keybert==0.9.0
MarkupSafe==2.1.1
requests==2.32.5
Authlib==1.6.6
psutil==7.2.1
